Change parser module name to phrase_parser
