+++ Consultoria Inteligencia Artificial/Projetos/Logica Quantificacional/phrase_parser/layers/abstract_layer.py
