+++ Consultoria Inteligencia Artificial/Projetos/Logica Quantificacional/phrase_parser/layers/existential_quantification_layer.py
@@ -1,4 +1,4 @@
-from parser.layers.abstract_layer import Layer
+from phrase_parser.layers.abstract_layer import Layer
 
 
 class ExistentialQuantificationLayer(Layer):
