+++ Consultoria Inteligencia Artificial/Projetos/Logica Quantificacional/phrase_parser/layers/simple_statement_layer.py
@@ -1,4 +1,4 @@
-from parser.layers.abstract_layer import Layer
+from phrase_parser.layers.abstract_layer import Layer
 
 
 class SimpleStatementLayer(Layer):
