+++ Consultoria Inteligencia Artificial/Projetos/Logica Quantificacional/phrase_parser/test/test_parser.py
@@ -1,6 +1,6 @@
 import unittest
 
-from parser.sentence_parser import SentenceParser
+from phrase_parser.sentence_parser import SentenceParser
 
 
 class TestSentenceParser(unittest.TestCase):
