+++ Consultoria Inteligencia Artificial/Projetos/Logica Quantificacional/usage.py
@@ -1,4 +1,4 @@
-from parser.sentence_parser import SentenceParser
+from phrase_parser.sentence_parser import SentenceParser
 
 parser = SentenceParser("Nenhum homem não é humano.")
 result = str(parser.parse())
